Rename FunctionalUtils.h to BaseClass.h
