+++ k4FWCore/include/k4FWCore/BaseClass.h
