+++ test/k4FWCoreTest/src/components/ExampleFunctionalTransformer.cpp
@@ -8,7 +8,7 @@
 #include "podio/CollectionBase.h"
 
 // Define BaseClass_t
-#include "k4FWCore/FunctionalUtils.h"
+#include "k4FWCore/BaseClass.h"
 
 #include <string>
