+++ test/k4FWCoreTest/src/components/ExampleFunctionalTransformerMultiple.cpp
@@ -10,7 +10,7 @@
 #include "podio/UserDataCollection.h"
 
 // Define BaseClass_t
-#include "k4FWCore/FunctionalUtils.h"
+#include "k4FWCore/BaseClass.h"
 
 #include <string>
